Amélioration de l'Interface Utilisateur et Gestion de l'Authentification

+++ src/app/pages/picture/detail-picture/detail-picture.component.ts
@@ -1,6 +1,9 @@
 import { Component, OnInit } from '@angular/core';
 import { ActivatedRoute, Router } from '@angular/router';
+import { Category } from 'src/app/mockData/category';
 import { Picture } from 'src/app/mockData/pictures';
+import { AuthenticationService } from 'src/app/security/services/authentication.service';
+import { CategoryService } from 'src/app/service/category.service';
 import { PicturesService } from 'src/app/service/pictures.service';
 
 @Component({
@@ -12,11 +15,16 @@ export class DetailPictureComponent implements OnInit{
 
 
   picturList: Picture [];
+  categoryList: Category [];
   picture: Picture | undefined;
+  isLoggedIn: boolean = false;
+
   constructor (
       private routeActive: ActivatedRoute,
       private route: Router ,
-      private pictureService : PicturesService
+      private pictureService : PicturesService,
+      private authService: AuthenticationService
+
     ){}
 
   ngOnInit() {
@@ -27,20 +35,27 @@ export class DetailPictureComponent implements OnInit{
       this.pictureService.getPictureById(+pictureId)
       .subscribe(picture => this.picture = picture);
     }
+    // Initialisez l'état de connexion
+    this.isLoggedIn = this.authService.isLoggedIn();
 
   }
 
   deletePicture(picture: Picture) {
-    alert(`Voulez-vous vraiment suppprimer ${picture.name} ?`);
-    this.pictureService.deletePictureById(picture.id)
-    .subscribe(() => this.goBack());
+    let categoryNames = picture.category.map(categ => categ.nameCate).join(', ');
+    let confirmation = confirm(`Voulez-vous vraiment supprimer cette photo de la catégorie ${categoryNames} ?`);
+    if (confirmation) {
+      this.pictureService.deletePictureById(picture.id)
+        .subscribe(() => this.goBack());
+    }
   }
-  goBack() {
-    this.route.navigate(['/pictures'])
-  }
+
 
   goToEditPicture(picture: Picture) {
     this.route.navigate(['/edit/picture', picture.id]);
+  }
+
+  goBack() {
+    this.route.navigate(['/pictures'])
   }
 
 }
